lib: rename da.ts to d.ts
It now deals with both `@da` and `@dr`, so should be named accordingly,
similar to r.ts.
