+++ src/d.ts
